MUD2 Resources, Utilities, Research

+++ packages/contracts/src/systems/IncrementSystem.sol
@@ -5,9 +5,9 @@ import { System } from "@latticexyz/world/src/System.sol";
 import { Counter } from "codegen/Tables.sol";
 
 contract IncrementSystem is System {
-  function increment() public returns (uint32) {
-    uint32 counter = Counter.get();
-    uint32 newValue = counter + 1;
+  function increment() public returns (uint256) {
+    uint256 counter = Counter.get();
+    uint256 newValue = counter + 1;
     Counter.set(newValue);
     return newValue;
   }
